adding names arg to logical type to_json

+++ lang/py/src/avro/io.py
@@ -302,7 +302,7 @@ class BinaryDecoder(object):
     return datetime.date(1970, 1, 1) + datetime.timedelta(days_since_epoc)
 
   def _build_time_object(self, value, unit_multiplier):
-    value = int(value * (1000000 // unit_multiplier))
+    value = value * (1000000 // unit_multiplier)
     value, microseconds =  value // 1000000, value % 1000000
     value, seconds = value // 60, value % 60
     value, minutes = value // 60, value % 60
